+++ src/main/java/com/daniel/coupons/logic/CompaniesLogic.java
@@ -18,25 +18,33 @@ public class CompaniesLogic {
     private ICompanyRepository companyRepository;
 
     public long createCompany(Company company) throws ApplicationException {
+        validateCreateCompany(company);
         validateCompany(company);
         CompanyEntity companyEntity = new CompanyEntity(company);
         companyRepository.save(companyEntity);
         return companyEntity.getId();
     }
 
-    public Company getCompanyById(long id){
+    public Company getCompanyById(long id) throws ApplicationException {
+        if (!isCompanyExistById(id)) {
+            throw new ApplicationException(ErrorType.COMPANY_DOESNT_EXIST);
+        }
         CompanyEntity companyEntity = companyRepository.findById(id).get();
         Company company = new Company(companyEntity);
         return company;
     }
 
     public void updateCompany(Company company) throws ApplicationException {
+        validateUpdateCompany(company);
         validateCompany(company);
         CompanyEntity companyEntity = new CompanyEntity(company);
         companyRepository.save(companyEntity);
     }
 
-    public void deleteCompanyById(long id){
+    public void deleteCompanyById(long id) throws ApplicationException {
+        if (!isCompanyExistById(id)) {
+            throw new ApplicationException(ErrorType.COMPANY_DOESNT_EXIST);
+        }
         companyRepository.deleteById(id);
     }
 
@@ -59,6 +67,21 @@ public class CompaniesLogic {
         return companyRepository.existsById(id);
     }
 
+    private void validateCreateCompany(Company company) throws ApplicationException {
+        if(companyRepository.existsByName(company.getName())){
+            throw new ApplicationException(ErrorType.COMPANY_ALREADY_EXIST);
+        }
+    }
+
+    private void validateUpdateCompany(Company company) throws ApplicationException {
+        if(!isCompanyExistById(company.getId())){
+            throw new ApplicationException(ErrorType.COMPANY_DOESNT_EXIST);
+        }
+        if(companyRepository.findById(company.getId()).get().getName() != company.getName()){
+            throw new ApplicationException(ErrorType.COMPANY_NAME_CANT_BE_CHANGED);
+        }
+    }
+
     private void validateCompany(Company company) throws ApplicationException {
         if(company.getName() == null || company.getName().isEmpty()){
             throw new ApplicationException(ErrorType.INVALID_COMPANY_NAME);
@@ -66,9 +89,5 @@ public class CompaniesLogic {
         if(company.getPhoneNumber() == null || company.getPhoneNumber().isEmpty()){
             throw new ApplicationException(ErrorType.INVALID_COMPANY_PHONE_NUMBER);
         }
-
-//        if(this.companiesDal.isCompanyExistByName(company.getName())){
-//            throw new ApplicationException(ErrorType.COMPANY_ALREADY_EXIST);
-//        }
     }
 }
